move config object generation to setup/config
